standalone/load_path: rename to init
