+++ Library/Homebrew/standalone.rb
@@ -3,5 +3,5 @@
 
 # This file should be the first `require` in all entrypoints outside the `brew` environment.
 
-require_relative "standalone/load_path"
+require_relative "standalone/init"
 require_relative "standalone/sorbet"
