+++ Library/Homebrew/startup.rb
@@ -3,7 +3,7 @@
 
 # This file should be the first `require` in all entrypoints of `brew`.
 
-require_relative "standalone/load_path"
+require_relative "standalone/init"
 require_relative "startup/ruby_path"
 require "startup/config"
 require_relative "startup/bootsnap"
